+++ PriorityQueue.h
@@ -11,10 +11,10 @@ struct Element
 {
 	unsigned int val;
 	unsigned int parent;
-	unsigned int p;
+	double p;
 
-	Element() : val{ 0 }, parent{ 0 }, p{ std::numeric_limits<unsigned int>::max() } {}
-	Element(unsigned int val, unsigned int parent, unsigned int p) : val{ val }, parent{ parent }, p{ p } {}
+	Element() : val{ 0 }, parent{ 0 }, p{ std::numeric_limits<double>::max() } {}
+	Element(unsigned int val, unsigned int parent, double p) : val{ val }, parent{ parent }, p{ p } {}
 };
 
 class PriorityQueue
@@ -25,17 +25,17 @@ private:
 
 	bool find_next(unsigned int start, int x, unsigned int &i);
 	void heapify(unsigned int i);
-	void inner_priority(unsigned int i, unsigned int p);
+	void inner_priority(unsigned int i, double p);
 
 public:
 	PriorityQueue();
 	~PriorityQueue();
 
-	void insert(int x, unsigned int parent, unsigned int p);
+	void insert(int x, unsigned int parent, double p);
 	bool empty();
 	unsigned int top();
 	unsigned int pop();
-	void priority(unsigned int x, unsigned int p);
+	void priority(unsigned int x, double p);
 	bool in_queue(unsigned int x);
 	bool in_queue(unsigned int x, unsigned int &i);
 	std::shared_ptr<Element> at(unsigned int i);
